requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python_dotenv
psycopg2-binary